Cleaning up init and index with module functions
