+++ app-main/public/js/main/avatars.js
