+++ app-main/public/js/main/events/eventHandlers.js
@@ -1,4 +1,4 @@
-import { quill } from "../firebase/firebase_quill_init.js";
+import { quill } from "../../firebase/firebase_quill_init.js";
 
 // cached document elements
 const overlay = document.querySelector(".overlay");
