+++ app-main/public/js/main/init.js
@@ -1,0 +1,16 @@
+import { quill } from "./quill_init.js";
+import setEventListeners from "./events/eventListeners.js";
+import setUser from "./setUser.js";
+import showContainers from "./showContainers.js";
+
+export default function init() {
+  quill.root.focus();
+  // Clear all containers of "hidden" attributes
+  showContainers();
+
+  // Set up User in Local Storage
+  setUser();
+
+  // initialize the event listeners;
+  setEventListeners();
+}
