+++ app-main/public/js/main/quill_init.js
